- Compiler error fix on Android - AndroidX fix - Restructure and cleaning - ApiKey toast message
