+++ example/lib/core/config.dart
