+++ example/lib/core/constants.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+
+const Color headlineColor = Color(0xFF061230);
+const Color textColor = Color(0xFF6D7486);
