+++ example/lib/core/utils.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+extension ListSpaceBetweenExtension on List<Widget> {
+  List<Widget> withSpaceBetween({final double? width, final double? height}) =>
+      [
+        for (int i = 0; i < length; i++) ...[
+          if (i > 0) SizedBox(width: width, height: height),
+          this[i],
+        ],
+      ];
+}
