+++ example/pubspec.yaml
@@ -29,6 +29,7 @@ dependencies:
   shared_preferences: 2.1.0
   http: 0.13.5
   provider: ^5.0.0
+  fluttertoast: ^8.2.4
   # geolocator: ^9.0.2
 
 dev_dependencies:
@@ -64,5 +65,5 @@ flutter:
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
 
-  fonts:
-  - family: HelveticaNeue
+  # fonts:
+  # - family: HelveticaNeue
